requirements.txt
neo4j==5.14.1
pandas==2.1.4
python-dotenv==1.0.0
streamlit==1.31.1
langgraph==0.0.19
langchain==0.1.4
langchain-openai==0.0.5
langchain-core==0.1.17
pydantic==2.5.3 